fix navbar mobile popover
